✨ 인터페이스 생성을 통한 OCP 준수, review 해결, 로그인 마지막에 사용자 정보 전달
