+++ src/main/java/com/goorm/wordsketch/util/OAuth2LoginSuccessHandler.java
@@ -1,6 +1,8 @@
 package com.goorm.wordsketch.util;
 
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.goorm.wordsketch.entity.User;
 import com.goorm.wordsketch.repository.UserRepository;
 import com.goorm.wordsketch.service.JwtService;
 import jakarta.servlet.ServletException;
@@ -12,6 +14,9 @@ import org.springframework.security.web.authentication.AuthenticationSuccessHand
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
+import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
@@ -19,13 +24,28 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
 
     private final JwtService jwtService;
 
+    private final UserRepository userRepository;
+
+    private final ObjectMapper objectMapper;
+
+    // Todo: 배포하면 url 변경 필요
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
         if (null != authentication) {
             jwtService.createAccessToken(response, authentication);
             jwtService.createRefreshToken(response, authentication);
 
-            response.sendRedirect("http://localhost:3000/");
+            Optional<User> optionalUser = userRepository.findByEmail(authentication.getName());
+            if (optionalUser.isPresent()) {
+                User user = optionalUser.get();
+                String userJson = objectMapper.writeValueAsString(user);
+                String encodedUserJson = URLEncoder.encode(userJson, StandardCharsets.UTF_8);
+                response.sendRedirect("http://localhost:3000/login/oauth2/userInfo?user=" + encodedUserJson);
+            } else {
+                String errorMessage = URLEncoder.encode("User not found", StandardCharsets.UTF_8);
+                response.sendRedirect("http://localhost:3000/login/oauth2/userInfo?error=" + errorMessage);
+            }
         }
+
     }
 }
